+++ src/Ext/Techno/Body.Internal.cpp
@@ -176,6 +176,30 @@ void TechnoExt::ExtData::InitializeDisplayInfo()
 	}
 }
 
+void TechnoExt::ExtData::InitializeUnitIdleAction()
+{
+	auto const pThis = this->OwnerObject();
+
+	if (pThis->WhatAmI() != AbstractType::Unit || !pThis->HasTurret())
+		return;
+
+	auto const pType = pThis->GetTechnoType();
+
+	if (this->TypeExtData->AutoFire || pType->TurretSpins)
+		return;
+
+	if (this->TypeExtData->UnitIdleRotateTurret.Get(RulesExt::Global()->UnitIdleRotateTurret))
+		this->UnitIdleAction = true;
+
+	this->UnitIdleTurretROT = static_cast<int>(pThis->SecondaryFacing.ROT.GetDir());
+
+	if (!SessionClass::IsSingleplayer())
+		return;
+
+	if (this->TypeExtData->UnitIdlePointToMouse.Get(RulesExt::Global()->UnitIdlePointToMouse))
+		this->UnitIdleActionSelected = true;
+}
+
 void TechnoExt::ExtData::InitializeAttachEffects()
 {
 	if (auto pTypeExt = this->TypeExtData)