+++ src/Ext/Techno/Hooks.cpp
@@ -76,6 +76,7 @@ DEFINE_HOOK(0x6F42F7, TechnoClass_Init, 0x2)
 	pExt->InitializeLaserTrails();
 	pExt->InitializeAttachEffects();
 	pExt->InitializeDisplayInfo();
+	pExt->InitializeUnitIdleAction();
 
 	return 0;
 }