+++ src/Ext/TechnoType/Body.h
@@ -231,6 +231,9 @@ public:
 		Nullable<AnimTypeClass*> Wake_Grapple;
 		Nullable<AnimTypeClass*> Wake_Sinking;
 
+		Nullable<bool> UnitIdleRotateTurret;
+		Nullable<bool> UnitIdlePointToMouse;
+
 		struct LaserTrailDataEntry
 		{
 			ValueableIdx<LaserTrailTypeClass> idxType;
@@ -457,6 +460,9 @@ public:
 			, Wake { }
 			, Wake_Grapple { }
 			, Wake_Sinking { }
+
+			, UnitIdleRotateTurret {}
+			, UnitIdlePointToMouse {}
 		{ }
 
 		virtual ~ExtData() = default;