+++ docs/Whats-New.md
@@ -444,6 +444,7 @@ New:
 - `RemoveDisguise` now works on vehicle disguises (by Trsdy)
 - Damage multiplier for different houses (by CrimRecya)
 - Display factories, superweapons, iron curtain and temporal progress on specific technos (by CrimRecya)
+- Units Rotate Turret When in Idle Action (by CrimRecya)
 - Allow anchoring extended tooltips to the left side of the sidebar (by Trsdy)
 - Toggle to allow spawned aircraft to attack immediately after being spawned (by Starkku)
 - `ZAdjust` for OverlayTypes (by Starkku)